Create crate-wide `iters` module
